Feat: config socket
